Split server files
